flashcards/app/api/pyproject.toml
# Furikaeri API - use `uv lock` to update lockfile, then `uv sync` to install.
[project]
name = "furikaeri-api"
version = "0.1.0"
description = "JLPT flashcards API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "python-multipart>=0.0.9",
    "PyYAML>=6.0",
    "fugashi[unidic-lite]>=1.5",
    "argon2-cffi>=23.1",
    "PyJWT>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "httpx>=0.26",
    "ruff>=0.3",
    "pip-audit>=2.7",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
